Fix examples to new Rootbeer API ThreadConfig(int,int,long)

+++ examples/matrixmultiplication2/src/at/illecker/rootbeer/examples/matrixmultiplication2/MatrixMultiplication2Kernel.java
@@ -202,7 +202,7 @@ public class MatrixMultiplication2Kernel implements Kernel {
     Context context = rootbeer.createDefaultContext();
     Stopwatch watch = new Stopwatch();
     watch.start();
-    rootbeer.run(kernel, new ThreadConfig(blockSize, gridSize, blockSize
+    rootbeer.run(kernel, new ThreadConfig(blockSize, gridSize, (long) blockSize
         * gridSize), context);
     watch.stop();
